+++ examples/tessellation.rs
@@ -25,7 +25,7 @@ fn main() {
         implement_vertex!(Vertex, position);
 
         glium::VertexBuffer::new(&display, 
-            vec![
+            &[
                 Vertex { position: [-0.5, -0.5] },
                 Vertex { position: [ 0.0,  0.5] },
                 Vertex { position: [ 0.5, -0.5] },
@@ -36,7 +36,7 @@ fn main() {
     // building the index buffer
     let index_buffer = glium::IndexBuffer::new(&display,
                                                PrimitiveType::Patches { vertices_per_patch: 3 },
-                                               vec![0u16, 1, 2]).unwrap();
+                                               &[0u16, 1, 2]).unwrap();
 
     // compiling shaders and linking them together
     let program = glium::Program::new(&display,